+++ release/ewoh-0.6.0-rc2/ewoh-spark-app/server/modules/workflow/workflow.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { WorkflowController } from './workflow.controller';
+import { WorkflowInstanceService } from './workflow-instance.service';
 import { WorkflowService } from './workflow.service';
 
 @Module({
   controllers: [WorkflowController],
-  providers: [WorkflowService],
-  exports: [WorkflowService],
+  providers: [WorkflowService, WorkflowInstanceService],
+  exports: [WorkflowService, WorkflowInstanceService],
 })
 export class WorkflowModule {}